requirements.txt
adafruit-ampy==1.1.0
Adafruit-Blinka==8.46.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.73.0
Adafruit-PureIO==1.1.11
annotated-types==0.7.0
anyio==4.4.0
binho-host-adapter==0.1.6
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpy_cross==1.23.0.post2
numpy==2.0.1
orjson==3.10.7
paho-mqtt==2.1.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
pyftdi==0.55.4
Pygments==2.18.0
pyserial==3.5
python-dotenv==1.0.1
python-multipart==0.0.9
pyusb==1.2.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
scipy==1.14.0
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.5.0
starlette==0.37.2
termcolor==2.4.0
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
